+++ webapp/javascript/components/TagsBar.tsx
@@ -2,7 +2,12 @@ import React, { useState, useEffect, useRef } from 'react';
 import Button from '@ui/Button';
 import 'react-dom';
 import { TagsState } from '@pyroscope/redux/reducers/continuous';
-import Dropdown, { SubMenu, MenuItem, FocusableItem } from '@ui/Dropdown';
+import Dropdown, {
+  SubMenu,
+  MenuItem,
+  FocusableItem,
+  MenuGroup,
+} from '@ui/Dropdown';
 import { Prism } from '../util/prism';
 import styles from './TagsBar.module.css';
 
@@ -236,7 +241,7 @@ function LabelsSubmenu({
         )}
       >
         {GetFilter(tag)}
-        {GetTagValues(tag, tagValues)}
+        <MenuGroup takeOverflow>{GetTagValues(tag, tagValues)}</MenuGroup>
       </SubMenu>
     );
   });